mondodb_user provider fails in replicaset
The mongodb_user provider seems to fail when it tries to evaluate a user on a secondary mongodb instance. Adding rs.slaveOk() to relevant methods solves the problem for our use cases, only tested on a 5 node cluster with no arbiters. Not sure how well this applies to standalone instances of mongodb, or arbiters.