do not use subdirectory for PID file
PID file is currently stored in /var/run/mongodb/
However, /var/run and /var/lock are mounted as tmpfs.

So if the subdirectory is not managed by the init script (which is not
the case in MongoDB packaging), the service won't start after a reboot,
since PID subdirectory is not created.

By changing the path of PID file to /var/run/mongod.pid fix the issue,
since there is no need to create a directory.

Signed-off-by: Emilien Macchi <emilien.macchi@enovance.com>