Fix server acceptance tests
The mongodb::server::ensure parameter is no longer sufficient to
describe the state of the package, config files, and service. This was
causing the acceptance tests to fail when trying to uninstall mongodb
because the manifest was attempting to ensure the service was running
while the package was uninstalled. This design should be reworked, but
in the mean time this commit adds the parameters necessary to
successfully uninstall mongodb. It also adds "stopped" as a valid value
for service_ensure to correspond with the valid values for the package
parameter.