add arbiter configuration
This commit adds the ability to configure a mongodb member as an arbiter. The
feature is very simplistic and will check for each member if it's an arbiter.
If the arbiter is changed after provision, the configuration will not change.
This is both as a safe guard but would also add on complexity.

An example is provided in `tests/replicaset.pp`:

    mongodb_replset{'rsmain':
      members => ['mongo1:27017', 'mongo2:27017', 'mongo3:27017' ],
      arbiter => 'mongo3:27017',
    }

The above configuration will set `mongo3:27017` as an arbiter.