Allow mongo client to connect on conf spec port
Commit 10d8f22323a4bda128661ebe8216dea2594013dd forces the port to be
set to 27107 if no port is speficied. Unfortunately, if the mongod
process is run in configsvr mode or shardsvr mode it listens to another
port by default. (configsvr => 27019, shardsvr => 27018).

This commit aims to make mongo client connect on the proper port based
on the configuration the mongod process has been started with.