Allow one to remove the service from the catalog
Allow a deployer to not have the service resource from the catalog.

This feature is useful when one manages the service through another
program, like pacemaker.

With a pacemaker setup, a user ensure a service is stopped, pacemaker is
in charge of starting it. At a subsequent run, puppet will make sure the
service is stopped again. To avoid that, this feature offer an option to
the user to not manage the resource at all.