Enable IPv6 in mongodb provider
When the MongoDB server is configured to listen to an IPv6 address,
the mongo client needs to use the "--ipv6" command-line switch to
connect using IPv6 [1].

It is safe to enable IPv6 for all commands, as this is just enabling
support. Using an IPv4 address when specifying --ipv6 works fine.

https://docs.mongodb.org/manual/reference/program/mongo/#cmdoption--ipv6