Set dbpath_fix to false by default
Under Xenial, the current default value of dbpath_fix set to true leads
to idempotency issues because the data files may be created with the
nogroup group.  This can lead to puppet repeatedly fixing permissions
which is not ideal. The files that are getting changed are set to 0600
so the group ownership does not help.  By default, we should be setting
this to false because we should not need to manage the user/group
permissions with the default install provided by distrobutions.