Add MongoDB Enterprise on Debian & Documentation cleanup (#374)
* Add support for MongoDB Enterprise with Debian
* Fix markdown for documentation

A non documented global variable 'use_enterprise_repo' existed to allow
the use of the Entreprise packages for RedHat.
This patch adds the option to Debian. It also tries to choose the right
APT key depending on the MongoDB version and adds the global variable to
the documentation.

* Clean up the sections in the documentation

Sections were not rendered correctly in GitHub. This patch adds a space
before the section titles in order to parse them correctly and render
them as expected.

* Fix ToC rendering on github

Some elements in the Table of Contents are incorrectly rendered under
github. This is due to unexpected spaces in the markdown code. Removing
them will allow to render the markdown correctly.