Ticket: https://tickets.puppetlabs.com/browse/MODULES-1754 Topic: After turnining on mongodb auth..subsequent runs fail
Although this is an old ticket the problem still exist in latest mongodb.
If you enable auth you need to store credentials for facts and to make adjustments to your setup.
This fix allows you to distribute mongorc.js manually if you do not want to store credentials in your puppet repository.