(maint) fix puppet 5 spec failure
for some reason, the test for the mongos service was choking on Puppet 5.0.0 but not Puppet 4.10.1. the config file resources for the mongos service used the 'before' param with value Service['mongos']. This does not appear to work in Puppet 5.0.0 since the service resource wasn't being created if mange_service was false.