Add a retry block when hosts are added to replset
There is a race condition here where, if the master has changed since
the replset host add will fail. This happens during scale out of nodes
or when master is being migrated.

Adding a retry block here will give the hosts a chance to join the
cluster while the master is coming up.