Handle non-existing config file
When using the RH MongoDB 3.4 SCL the config file is in
/etc/opt/rh/rh-mongodb34/mongod.conf which causes this fact to fail
because the -syspaths package still installs the binaries to /usr/bin.
Since we don't use this fact, it's enough for now to gracefully handle
this.

Ideally we could use the value for mongodb::server::config but I'm not
aware of a way to do this in facts.