Describe disabling logpath to make syslog work
I was confused about how to make the `syslog` config work because, even though I had not set `logpath`, the default value kept `syslog` from working. I tried setting it to an empty string, but that failed with a failed pattern check.

I had to look into the code to find out that I can also set it to `false` and that made it work.