Connect to localhost in provider when possible
The provider previously used the external IP to connect instead of the
loopback.  When using the loopback, the option
`enableLocalhostAuthBypass` (true by default) allows to bypass
authentication when a user doesn't yet exists.  Which is exactly what
would prevent us from executing administrative tasks (like setting up
the replicaset) when auth is enabled, but not yet set up.

We can only connect to localhost when MongoDB is listening on either the
'bind all' address (0.0.0.0 or ::0) or the local loopback address
(127.0.0.1 or ::1).