Use a more suitable method of detecting if authentication is needed
The `db.serverCmdLineOpts()` method always requires authentication when
auth is enabled. It doesn't take the `enableLocalAuthBypass` parameter
into account, the subsequent authentication attempt always fails if a
user doesn't exists. This commit switches to the `rs.status()` method,
as that method does take the `enableLocalAuthBypass` parameter into
account and allows us to obtain a shell before with authentication
enabled, but before the user exists.