Users should be created before the database
Databases in MongoDB are created lazy, meaning there is no `create`
command. Thus if a record is inserted the DB is created.  When auth is
enabled, any attempt to insert data is denied because of missing auth.
However the `db.createUser()` function works because of the config
option `enableBypassLocalAuth`. For an overview see
https://www.mongodb.com/docs/manual/tutorial/deploy-replica-set-with-keyfile-access-control/