Remove obsolete file
The `.nodeset.yml` file only contains some old OS versions and was never
touched after it initial commit in 2013
(cce6136937a5dae98762ffb5bdc5c7aa4ea20276). This dates back from when
this module was maintained by PuppetLabs.

The entry in the README.md file was never correct anyway, linking to the
proper file, but calling it an XML  file.